refactor: renamed package, finance link now returns new page
