+++ src/main/java/yahoo/YahooFinance.java
@@ -1,0 +1,7 @@
+package yahoo;
+
+import base.BasePage;
+
+public class YahooFinance extends BasePage {
+    
+}
